api/requirements.txt
Django==4.0.5
asgiref==3.5.2
autopep8==1.6.0
dataclasses==0.6
django-cors-headers==3.13.0
django-stubs==1.12.0
djangorestframework-types==0.7.0
djangorestframework==3.13.1
flake8-isort==4.1.1
google-auth==2.8.0
isort==5.10.1
pycodestyle==2.8.0
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-order==1.0.1
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
toml==0.10.2
google-auth-oauthlib==0.5.2
